+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-## 1.0.0 (2024-10-25)
+## 1.0.2 (2024-10-31)
 
-- Feat: support remove `false` condition and if
+- Feat: support remove condition and if of booleanLiteral `true` and `const isX = true`
+
+## 1.0.1 (2024-10-25)
+
+- Feat: support remove booleanLiteral `false` condition and if
 
 ## 1.0.0 (2024-10-25)